requirements.txt
django==5.1.1
psycopg2==2.9.9
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
django-allauth==64.2.1
django-debug-toolbar==4.4.6
pillow==10.4.0